consumer/requirements.txt
azure-eventhub
aiohttp
azure-eventhub-checkpointstoreblob-aio
azure-identity